Refactoried virtual host support.
Merged setApplicationNames into setVirtualHostRoot

Got rid of the _vh_trunc variable. Truncation now happens immediately,
which simplifies the implementation.

Added the shiftNameToApplication method, which allows side-effect path
elements (e.g. ++skin++ZopeTop) to be moved into the "application"
part of the url, so that we son't have to keep track of them in
context wrappers.