*Finally* renamed setViewSkin to setPresentationSkin.
Merged setApplicationNames into setVirtualHostRoot.

Added the shiftNameToApplication method, which allows side-effect path
elements (e.g. ++skin++ZopeTop) to be moved into the "application"
part of the url, so that we son't have to keep track of them in
context wrappers.

Changesd assertions to not use _vh_trunc, which is now gone.

Fixed a test that wasn't consistent with the definition of setVirtualHostRoot.