Created a more egeneric zope-wsgi implementation that works woth other publisher applications.  This is based on the wsgi support created by Eckart Hertzler in zope.app.wsgi.
Eventually (somewhat soon), the zope.app version will be modified to
use this one, or will be replaced by this one. First, some refactoring
of zope.app's publication will need to be refactored.

This implemenattion is far more compicated that I think it should be,
due to the wanton complexity of the HTTP response implementation. This
really needs another look.