Fixed a bug in the publisher reported by Roger.
When the base is inserted into the body, the body is already encoded t a 
string. If the base is a unicode string, then the entire body used to end 
up as a unicode string. This bug was fixed by applying the correct 
encoding on the base URL, if it is a unicode string.